[REF] store: replace ConnectedComponent by useStore hook
Closes #304
