+++ doc/readme.md
@@ -10,6 +10,7 @@ owl
     Component
     Context
     QWeb
+    Store
     useState
     core
         EventBus
@@ -23,13 +24,13 @@ owl
         useState
         useRef
         useSubEnv
+        useStore
+        useDispatch
+        useGetters
     router
         Link
         RouteComponent
         Router
-    store
-        Store
-        ConnectedComponent
     tags
         xml
     utils
@@ -39,6 +40,7 @@ owl
         loadTemplates
         whenReady
 ```
+
 
 Note that for convenience, the `useState` hook is also exported at the root of the `owl` object.
